+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 
 ## v1.10.0
 
-- Add metadata to TopicPartition type and commit() (#1410).
+ - Added metadata to TopicPartition type and commit() (#1410).
+ - Added consumer.memberid() for getting member id assigned to 
+   the consumer in a consumer group (#1154).
 
 
 ## v1.9.2